test: split PDF document structure cases
